adds search bar to docs, moves document related views to own file
